% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_edges.R
\name{tm_edges}
\alias{tm_edges}
\alias{opt_tm_edges}
\title{Map layer: edges of a (sf)network}
\usage{
tm_edges(
  col = tmap::tm_const(),
  col.scale = tmap::tm_scale(),
  col.legend = tmap::tm_legend(),
  col.free = NA,
  lwd = tmap::tm_const(),
  lwd.scale = tmap::tm_scale(),
  lwd.legend = tmap::tm_legend(),
  lwd.free = NA,
  lty = tmap::tm_const(),
  lty.scale = tmap::tm_scale(),
  lty.legend = tmap::tm_legend(),
  lty.free = NA,
  col_alpha = tmap::tm_const(),
  col_alpha.scale = tmap::tm_scale(),
  col_alpha.legend = tmap::tm_legend(),
  col_alpha.free = NA,
  from = 0,
  to = 1,
  linejoin = "round",
  lineend = "round",
  plot.order = tmap::tm_plot_order("lwd", reverse = TRUE, na.order = "bottom"),
  zindex = NA,
  group = NA,
  group.control = "check",
  popup = tmap::tm_popup(),
  popup.vars = NA,
  popup.format = list(),
  hover = NA,
  id = "",
  options = opt_tm_edges(),
  ...
)

opt_tm_edges(lines.only = "yes", offset_start = 0, offset_end = 0)
}
\arguments{
\item{col, col.scale, col.legend, col.free}{Visual variable that determines the col color. See details.}

\item{lwd, lwd.scale, lwd.legend, lwd.free}{Visual variable that determines the line width. See details.}

\item{lty, lty.scale, lty.legend, lty.free}{Visual variable that determines the line type. See details.}

\item{col_alpha, col_alpha.scale, col_alpha.legend, col_alpha.free}{Visual variable that determines the border color alpha transparency. See details.}

\item{from, to}{Numbers between 0 and 1 (where \code{to >= from}) that indicate which part of each edge is drawn. By default full lines, so \code{from} and \code{to} are respectively 0 and 1.}

\item{linejoin, lineend}{line join and line end. See \code{\link[grid:gpar]{gpar}} for details.}

\item{plot.order}{Specification in which order the spatial features are drawn. See \code{tmap::tm_plot_order} for details.}

\item{zindex}{Map layers are drawn on top of each other. The \code{zindex} numbers (one for each map layer) determines the stacking order. By default the map layers are drawn in the order they are called.}

\item{group}{Name of the group to which this layer belongs. This is only relevant in view mode, where layer groups can be switched (see \code{group.control})}

\item{group.control}{In view mode, the group control determines how layer groups can be switched on and off. Options: \code{"radio"} for radio buttons (meaning only one group can be shown), \code{"check"} for check boxes (so multiple groups can be shown), and \code{"none"} for no control (the group cannot be (de)selected).}

\item{popup}{popup specification for \code{"view"} mode, the output of \code{\link[tmap:tm_popup]{tmap::tm_popup()}}.
It determines the data variables shown in the popup table, the popup title,
and the popup layout. This replaces the deprecated arguments \code{popup.vars}
and \code{popup.format}.}

\item{popup.vars}{(Deprecated.) Use \code{popup} with \code{\link[tmap:tm_popup]{tmap::tm_popup()}} instead
(via its \code{vars} argument). Names of data variables that are shown in the
popups in \code{"view"} mode. Set \code{popup.vars} to \code{TRUE} to show all variables in
the shape object. Set \code{popup.vars} to \code{FALSE} to disable popups. Set
\code{popup.vars} to a character vector of variable names to show those variables
in the popups. The default (\code{NA}) depends on whether visual variables
(e.g. \code{col}) are used. If so, only those are shown. If not, all variables in
the shape object are shown.}

\item{popup.format}{(Deprecated.) Use \code{popup} with \code{\link[tmap:tm_popup]{tmap::tm_popup()}} instead
(via its \code{format} argument). List of formatting options for the popup values.
See the argument \code{legend.format} for options. Only applicable for numeric
data variables. If one list of formatting options is provided, it is applied
to all numeric variables of \code{popup.vars}. Also, a (named) list of lists can
be provided. In that case, each list of formatting options is applied to the
named variable.}

\item{hover}{name of the data variable that specifies the hover labels (view mode only). Set to \code{FALSE} to disable hover labels. By default \code{FALSE}, unless \code{id} is specified. In that case, it is set to \code{id},}

\item{id}{name of the data variable that specifies the indices of the spatial
features. Only used for \code{"view"} mode.}

\item{options}{options passed on to the corresponding \verb{opt_<layer_function>} function}

\item{...}{passed on to \code{\link[tmap:tm_lines]{tmap::tm_lines()}}.}

\item{lines.only}{should only line geometries of the shape object (defined in \code{\link[tmap:tm_shape]{tmap::tm_shape()}}) be plotted, or also other geometry types (like polygons)? By default \code{"ifany"}, which means \code{TRUE} in case a geometry collection is specified.}

\item{offset_start, offset_end}{Offset in coordinates (usually meters) of the start and end points.}
}
\value{
a \link[tmap:tmap-element]{tmap::tmap-element}, supposed to be stacked after \code{\link[tmap:tm_shape]{tmap::tm_shape()}} using the \code{+} operator. The \verb{opt_<layer_function>} function returns a list that should be passed on to the \code{options} argument.
}
\description{
Map layer that draws the edges of a (sf)network.
}
\examples{
library(tmap)
library(sfnetworks)

sfn = as_sfnetwork(roxel)

tm_shape(sfn) +
	tm_network()
	
tm_shape(sfn) +
	tm_edges(col = "type", lwd = 4) +
	tm_nodes()


tm_shape(sfn) +
	tm_edges(col = "type", lwd = 4, from = 0.3, to = 0.4) +
	tm_nodes()
}
