% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_network.R
\name{tm_network}
\alias{tm_network}
\title{Map layer: (sf)network}
\usage{
tm_network()
}
\value{
a \link[tmap:tmap-element]{tmap::tmap-element}, supposed to be stacked after \code{\link[tmap:tm_shape]{tmap::tm_shape()}} using the \code{+} operator. The \verb{opt_<layer_function>} function returns a list that should be passed on to the \code{options} argument.
}
\description{
Map layer that draws a network. For more (total) flexibility, please use \code{\link{tm_edges}} and \code{\link{tm_nodes}}.
}
\examples{
library(tmap)
library(sfnetworks)

sfn = as_sfnetwork(roxel)

tm_shape(sfn) +
	tm_network()
	
tm_shape(sfn) +
	tm_edges(col = "type", lwd = 4) +
	tm_nodes()


tm_shape(sfn) +
	tm_edges(col = "type", lwd = 4, from = 0.3, to = 0.4) +
	tm_nodes()
}
