% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tmap.networks-package}
\alias{tmap.networks}
\alias{tmap.networks-package}
\title{Extension for tmap: networks}
\description{
Networks from sfnetworks are supported and several network specifc layer functions are added
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-tmap/tmap.networks}
  \item \url{https://r-tmap.github.io/tmap.networks/}
  \item Report bugs at \url{https://github.com/r-tmap/tmap.networks/issues}
}

}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\concept{GIS}
\concept{bubble map}
\concept{choropleth}
\concept{statistical maps}
\concept{thematic maps}
