% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_linnet.R
\name{tmapGetShapeMeta1.linnet}
\alias{tmapGetShapeMeta1.linnet}
\title{Implementation of tmapGetShapeMeta1 for linnet}
\usage{
\method{tmapGetShapeMeta1}{linnet}(shp, o)
}
\value{
list used by tmap internally
}
\description{
Implementation of tmapGetShapeMeta1 for linnet
}
\keyword{internal}
