% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_linnet.R
\name{tmapGetShapeMeta2.linnet}
\alias{tmapGetShapeMeta2.linnet}
\title{Implementation of tmapGetShapeMeta2 for linnet}
\usage{
\method{tmapGetShapeMeta2}{linnet}(shp, smeta, o)
}
\value{
list used by tmap internally
}
\description{
Implementation of tmapGetShapeMeta2 for linnet
}
\keyword{internal}
