% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_linnet.R
\name{tmapSubsetShp.linnet}
\alias{tmapSubsetShp.linnet}
\title{Implementation of tmapSubsetShp for linnet}
\usage{
\method{tmapSubsetShp}{linnet}(shp, vars)
}
\value{
list used by tmap internally
}
\description{
Implementation of tmapSubsetShp for linnet
}
\keyword{internal}
